@@ -27,19 +27,19 @@
     "dayjs": "^1.11.7",
     "draft-js": "^0.11.7",
     "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
+    "eslint-config-next": "^14.0.1",
     "framer-motion": "^10.12.4",
     "html-react-parser": "^3.0.16",
     "multer": "^1.4.5-lts.1",
-    "next": "13.3.0",
+    "next": "^14.0.1",
     "next-connect": "^1.0.0",
     "next-share": "^0.20.0",
     "nookies": "^2.5.2",
     "postcss": "8.4.23",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-cookie-consent": "^8.0.1",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.9",
     "react-icons": "^4.8.0",
     "react-input-mask": "^2.0.4",
